+++ packages/app-template-default/client/plugins.ts
@@ -9,6 +9,7 @@ import install from '@nocobase/app-plugin-install/client';
 import notificationProvider from '@nocobase/app-plugin-notification-provider/client';
 import routesExample from '@nocobase/app-plugin-routes-example/client';
 import file from '@nocobase/app-plugin-file/client';
+import i18n from '@nocobase/app-plugin-i18n/client';
 import workflow from '@nocobase/app-plugin-workflow/client';
 import notification from '@nocobase/app-plugin-notification/client';
 import systemInfo from '@nocobase/app-plugin-system-info/client';
@@ -19,6 +20,7 @@ const clientPlugins: AppClientPlugins = defineClientPlugins([
   authentication(),
   authorization(),
   dataProvider(),
+  i18n(),
   install(),
   notificationProvider(),
   routesExample(),